requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2